Make Span abstract
